chore: split the nexus and volume parts of the volume service
Given that the nexus is a "low level" api, split it from the volume api
