+++ control-plane/agents/core/src/nexus/registry.rs
